Catalog  MergeConfilts resolved
